isi_mcp_demo/isi_mcp/requirements.txt
fastmcp<3.0.0
uvicorn
isilon-sdk
ansible-core
ansible-runner
jinja2
packaging
pyyaml
pytest
httpx
